requirements.txt
flask
boto3
tqdm
python-dotenv
requests
Pillow